Verifying amount of returns, Testing correct verification
